more refactoring + create utility fn
