+++ integration-tests/tests/upsert.rs
@@ -17,7 +17,7 @@ async fn upsert() -> TestResult {
         .user()
         .upsert(
             user::id::equals(user_id.to_string()),
-            (
+            user::create(
                 "Brendan".to_string(),
                 vec![user::id::set(user_id.to_string())],
             ),
@@ -33,7 +33,7 @@ async fn upsert() -> TestResult {
         .user()
         .upsert(
             user::id::equals(user_id.to_string()),
-            (
+            user::create(
                 "Oscar".to_string(),
                 vec![user::id::set(user_id.to_string())],
             ),
@@ -41,6 +41,7 @@ async fn upsert() -> TestResult {
         )
         .exec()
         .await?;
+
     assert_eq!(user.id, user_id);
     assert_eq!(user.name, "Oscar");
